+++ source/tv/phantombot/scripts/core/Moderation.java
@@ -14,14 +14,21 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package tv.phantombot.scripts.core;
 
-import java.util.HashMap;
-import java.util.regex.Pattern;
-import java.util.regex.Matcher;
+import net.engio.mbassy.listener.Handler;
 
-import org.json.JSONObject;
-import org.json.JSONArray;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+
+import tv.phantombot.PhantomBot;
+import tv.phantombot.event.Listener;
+import tv.phantombot.event.irc.message.IrcModerationEvent;
+import tv.phantombot.scripts.core.ModerationUtil.ChatMessage;
 
 /**
  * This system has a lot of repeating code, mostly because when matching spam, 
@@ -29,18 +36,40 @@ import org.json.JSONArray;
  * This doesn't allow us to make functions to get "totals" of things and reuse them.
  * 
  * @author ScaniaTV
-// */
-public final class Moderation {
+ */
+public class Moderation extends ModerationUtil implements Listener {
     private static final Moderation INSTANCE = new Moderation();
-    private final Pattern URL_PATTERN = Pattern.compile("((?:(http|https|rtsp):\\\\/\\\\/(?:(?:[a-z0-9\\\\$\\\\-\\\\_\\\\.\\\\+\\\\!\\\\*\\\\\\'\\\\(\\\\)\\\\,\\\\;\\\\?\\\\&\\\\=]|(?:\\\\%[a-fA-F0-9]{2})){1,64}(?:\\\\:(?:[a-z0-9\\\\$\\\\-\\\\_\\\\.\\\\+\\\\!\\\\*\\\\\\'\\\\(\\\\)\\\\,\\\\;\\\\?\\\\&\\\\=]|(?:\\\\%[a-fA-F0-9]{2})){1,25})?\\\\@)?)?((?:(?:[a-z0-9][a-z0-9\\\\-]{0,64}\\\\.)+(?:(?:aero|a[cdefgilmnoqrstuwxz])|(?:biz|b[abdefghijmnorstvwyz])|(?:com|c[acdfghiklmnoruvxyz])|d[ejkmoz]|(?:edu|e[cegrstu])|(?:fyi|f[ijkmor])|(?:gov|g[abdefghilmnpqrstuwy])|(?:how|h[kmnrtu])|(?:info|i[delmnoqrst])|(?:jobs|j[emop])|k[eghimnrwyz]|l[abcikrstuvy]|(?:mil|mobi|moe|m[acdeghklmnopqrstuvwxyz])|(?:name|net|n[acefgilopruz])|(?:org|om)|(?:pro|p[aefghklmnrstwy])|qa|(?:r[eouw])|(?:s[abcdeghijklmnortuvyz])|(?:t[cdfghjklmnoprtvwz])|u[agkmsyz]|(?:vote|v[ceginu])|(?:xxx)|(?:watch|w[fs])|y[etu]|z[amw]))|(?:(?:25[0-5]|2[0-4][0-9]|[0-1][0-9]{2}|[1-9][0-9]|[1-9])\\\\.(?:25[0-5]|2[0-4][0-9]|[0-1][0-9]{2}|[1-9][0-9]|[1-9]|0)\\\\.(?:25[0-5]|2[0-4][0-9]|[0-1][0-9]{2}|[1-9][0-9]|[1-9]|0)\\\\.(?:25[0-5]|2[0-4][0-9]|[0-1][0-9]{2}|[1-9][0-9]|[0-9])))(?:\\\\:\\\\d{1,5})?)(\\\\/(?:(?:[a-z0-9\\\\;\\\\/\\\\?\\\\:\\\\@\\\\&\\\\=\\\\#\\\\~\\\\-\\\\.\\\\+\\\\!\\\\*\\\\\\'\\\\(\\\\)\\\\,\\\\_])|(?:\\\\%[a-fA-F0-9]{2}))*)?(?:\\\\b|$)|(\\\\.[a-z]+\\\\/|magnet:\\/\\/|mailto:\\/\\/|ed2k:\\/\\/|irc:\\/\\/|ircs:\\/\\/|skype:\\/\\/|ymsgr:\\/\\/|xfire:\\/\\/|steam:\\/\\/|aim:\\/\\/|spotify:\\/\\/)");
-    private final Pattern YOUTUBE_PATTERN = Pattern.compile("!\\w{1,9}\\s((http(s)?:\\/\\/)?youtu(\\.be)?(be\\.com))");
-    private final Pattern URL_DECIPHER = Pattern.compile("(\\s?\\(?(dot|\\.){1,2}\\)?\\s?)");
+    private final ScheduledExecutorService chatCacheHandler = Executors.newSingleThreadScheduledExecutor();
+    public final ConcurrentHashMap<String, Object> settings = new ConcurrentHashMap<>();
+    public final CopyOnWriteArrayList<ChatMessage> chatCache = new CopyOnWriteArrayList<>();
     
     /**
      * Class constructor.
      */
-    private Moderation() {
+    public Moderation() {
+        //chatCacheHandler.schedule(this::cleanChatCache, 60, TimeUnit.SECONDS);
         
+        // Load all settings from the DB.
+        /*String[] keys = PhantomBot.instance().getDataStore().GetKeyList("chatModerator_Beta", "");
+        for (String key : keys) {
+            String value = PhantomBot.instance().getDataStore().get("chatModerator_Beta", key);
+            
+            if (key.startsWith("bool_")) {
+                // boolean
+                settings.put(key, Boolean.parseBoolean(value));
+            } else if (key.startsWith("int_")) {
+                // Int.
+                settings.put(key, Integer.parseInt(value));
+            } else if (key.startsWith("float_")) {
+                // Float.
+                settings.put(key, Float.parseFloat(value));
+            } else {
+                // String.
+                settings.put(key, value);
+            }
+        }*/
+        
+        settings.put("float_maxCapsPercent", Float.parseFloat("0.10"));
     }
     
     /**
@@ -53,320 +82,60 @@ public final class Moderation {
     }
     
     /**
-     * Method that checks if the message has a URL.
+     * Method that gets the setting for a moderation filter.
      * 
-     * @param message
-     * @param isDecipher
-     * @param isSongrequestsEnabled
+     * @param key
+     * @param defaultValue
      * @return 
      */
-    public boolean hasURL(String message, boolean isDecipher, boolean isSongrequestsEnabled) {
-        // If fake links like "google dot com", "google(.)com" should be changed to "google.com".
-        if (isDecipher) {
-            message = getDecipheredURLFromMessage(message);
+    private Object getSetting(String key, String defaultValue) {
+        Object returnValue = defaultValue;
+        
+        if (settings.containsKey(key)) {
+            returnValue = settings.get(key);
         }
         
-        boolean hasMatch = URL_PATTERN.matcher(message).find();
-        
-        if (isSongrequestsEnabled && YOUTUBE_PATTERN.matcher(message).find()) {
-            hasMatch = false;
-        }
-        
-        return hasMatch;
+        return returnValue;
     }
     
     /**
-     * Method that deciphers a URL from a message, if a link has "(dot)" and not a "." it gets replaced with a ".".
+     * Method that gets chat messages and saves them.
      * 
-     * @param message
-     * @return 
+     * @param event 
      */
-    private String getDecipheredURLFromMessage(String message) {
-        Matcher matches = URL_DECIPHER.matcher(message);
-        StringBuffer sb = null;
+    @Handler
+    public void onIrcModerationMessage(IrcModerationEvent event) {
+        /*chatCache.add(new ChatMessage(event.getSender(), event.getMessage()));
         
-        // Create a new string buffer if needed.
-        // The while doesn't need to be in here, but what the heck.
-        if (matches.find()) {
-            // Build our string buffer
-            sb = new StringBuffer(message.length());
-            // Replace the first match.
-            matches.appendReplacement(sb, ".");
-            
-            // Get all the other macthes and replace them.
-            while (matches.find()) {
-                matches.appendReplacement(sb, ".");
-            }
-            matches.appendTail(sb);
+        if (hasURL(event.getMessage(), true, false)) {
+            event.getSession().sayNow(".timeout " + event.getSender() + " 10");
+            return;
         }
-        
-        return (sb == null ? message : sb.toString());
+        if (hasSpecialLetters(event.getMessage(), (float)0.10)) {
+            event.getSession().sayNow(".timeout " + event.getSender() + " 10");
+            return;
+        }
+        if (hasMaximumCaps(event.getMessage(), 10, (float)settings.get("float_maxCapsPercent"), event.getTags().get("emotes"))) {
+            event.getSession().sayNow(".timeout " + event.getSender() + " 10");
+        }*/
     }
     
     /**
-     * Method that checks if a message or username has a blacklist.
-     * 
-     * @param username
-     * @param message
-     * @param blacklist
-     * @return 
+     * Method that handles clearing the cache.
      */
-    public boolean hasBlacklist(String username, String message, JSONObject blacklist) {
-        boolean hasBlacklist = false;
+    private void cleanChatCache() {
+        int maxChatCacheSize = PhantomBot.instance().getDataStore().GetInteger("chatModeration", "", "maxChatCacheSize");
+        int keepMessageTime = (PhantomBot.instance().getDataStore().GetInteger("chatModeration", "", "oneManSpamKeepTime") * 1000);
+        long currentTime = System.currentTimeMillis();
         
-        if (blacklist.has("_total") && blacklist.getInt("_total") > 0) {
-            JSONArray list = blacklist.getJSONArray("list");
-            
-            for (int i = 0; i < list.length(); i++) {
-                JSONObject wordObj = list.getJSONObject(i);
-                
-                // Handle the things.
-            }
-        }
-        
-        return hasBlacklist;
-    }
-    
-    /**
-     * Method that says if a string has too many non alpha chars.
-     * 
-     * @param message
-     * @param maxNonAlpha
-     * @param maxPercent
-     * @return 
-     */
-    public boolean hasMaximumNonAlphanumeric(String message, int maxNonAlpha, float maxPercent) {
-        boolean hasMaximumNonAlpha = false;
-        int messageLength = message.length();
-        int totalNonAlphanumericCount = 1;
-        
-        for (int i = 0; i < messageLength; i++) {
-            char c = message.charAt(i);
-            
-            if (!(Character.isLetterOrDigit(c) || Character.isWhitespace(c))) {
-                totalNonAlphanumericCount++;
-                if (totalNonAlphanumericCount >= maxNonAlpha || ((totalNonAlphanumericCount / messageLength) >= maxPercent)) {
-                    hasMaximumNonAlpha = true;
-                    break;
-                }
-            }
-        }
-        
-        return hasMaximumNonAlpha;
-    }
-    
-    /**
-     * Method that checks if multiple non-alphanumeric characters (random ones) are repeating multiple times in a row.
-     * 
-     * @param message
-     * @param maxLength
-     * @return 
-     */
-    public boolean hasLongNonAlphanumericSequence(String message, int maxLength) {
-        boolean hasLongRepeatingSequence = false;
-        int totalRepeatingSequence = 1;
-        
-        for (int i = 0; i < message.length(); i++) {
-            char c = message.charAt(i);
-            
-            if (!(Character.isLetterOrDigit(c) || Character.isWhitespace(c))) {
-                totalRepeatingSequence++;
-                if (totalRepeatingSequence >= maxLength) {
-                    hasLongRepeatingSequence = true;
-                    break;
-                }
+        for (int i = (chatCache.size() - 1); i >= 0; i--) {
+            if (i > maxChatCacheSize || (currentTime - chatCache.get(i).getTime()) < keepMessageTime) {
+                chatCache.remove(i);
             } else {
-                totalRepeatingSequence = 1;
+                // Break once the conditions don't match anymore
+                // No need to keep looping for nothing.
+                break;
             }
         }
-        
-        return hasLongRepeatingSequence;
-    }
-    
-    /**
-     * Method that checks if a character is repeated multiple times in a row.
-     * 
-     * @param message
-     * @param maxLength
-     * @return 
-     */
-    public boolean hasLongCharacterSequence(String message, int maxLength) {
-        boolean hasLongRepeatingSequence = false;
-        int totalRepeatingSequence = 1;
-        char lastCharacter = message.charAt(0);
-        
-        for (int i = 1; i < message.length(); i++) {
-            char c = message.charAt(i);
-            
-            if (c == lastCharacter) {
-                totalRepeatingSequence++;
-                if (totalRepeatingSequence >= maxLength) {
-                    hasLongRepeatingSequence = true;
-                    break;
-                }
-            } else {
-                totalRepeatingSequence = 1;
-            }
-            lastCharacter = c;
-        }
-        
-        return hasLongRepeatingSequence;
-    }
-    
-    /**
-     * Method that checks if a word is repeated multiple times in a row or in a phrase
-     * 
-     * @param message
-     * @param maxLength How many times a word can be repeated after another in a row.
-     * @param maxTimes How many times total in a message a word can be repeated.
-     * @param rawEmoteIndexes
-     * @return 
-     */
-    public boolean hasRepeatingWordsOrSequence(String message, int maxLength, int maxTimes, String rawEmoteIndexes) {
-        boolean hasLongRepeatingSequence = false;     
-        int totalRepeatingSequence = 1;
-        HashMap<String, Integer> repeats = new HashMap<>();
-        String[] messageParts = getMessageWithoutEmotes(message, rawEmoteIndexes).split(" ");
-        String lastWord = messageParts[0];
-
-        for (int i = 1; i < messageParts.length; i++) {
-            String word = messageParts[i];
-            
-            if (word.equalsIgnoreCase(lastWord)) {
-                totalRepeatingSequence++;
-                if (totalRepeatingSequence >= maxLength) {
-                    hasLongRepeatingSequence = true;
-                    break;
-                }
-            } else {
-                totalRepeatingSequence = 1;
-            }
-            
-            if (!repeats.containsKey(word)) {
-                repeats.put(word, 1);
-            } else {
-                int times = (repeats.get(word) + 1);
-                // If we said it the max times, break and return.
-                if (times >= maxTimes) {
-                   hasLongRepeatingSequence = true;
-                   break;
-                }
-                repeats.put(word, times);
-            }
-            lastWord = word;
-        }
-        
-        return hasLongRepeatingSequence;
-    }
-    
-    /**
-     * Method that gets a map of the emote indexes.
-     * 
-     * @param rawEmoteIndexes
-     * @return 
-     */
-    private HashMap<Integer, Integer> getEmotesIndexMap(String rawEmoteIndexes) {
-        HashMap<Integer, Integer> emoteIndexMap = new HashMap<>();
-        String[] indexArray;
-        
-        if (rawEmoteIndexes.length() > 0) {
-            indexArray = rawEmoteIndexes.split("/");
-        
-            for (int i = 0; i < indexArray.length; i++) {
-                String[] indexes = indexArray[i].substring(indexArray[i].indexOf(":") + 1).split(",");
-                for (int j = 0; j < indexes.length; j++) {
-                    String[] index = indexes[j].split("-");
-                
-                    emoteIndexMap.put(Integer.parseInt(index[0]), Integer.parseInt(index[1]));
-                }
-            }
-        }
-        
-        return emoteIndexMap;
-    }
-    
-    /**
-     * Method that removes all emotes from a message.
-     * 
-     * @param message
-     * @param rawEmoteIndexes
-     * @return 
-     */
-    private String getMessageWithoutEmotes(String message, String rawEmoteIndexes) {
-        HashMap<Integer, Integer> emoteIndexMap = getEmotesIndexMap(rawEmoteIndexes);
-        
-        for (int i = message.length() - 1; i >= 0; i--) {
-            if (emoteIndexMap.containsKey(i)) {
-                message = message.substring(0, i) + message.substring(emoteIndexMap.get(i) + 1);
-            }
-        }
-
-        return message.trim();
-    }
-    
-    /**
-     * Method that says if a message has too many caps.
-     * 
-     * @param message
-     * @param maxCaps
-     * @param maxPercent
-     * @param rawEmoteIndexes
-     * @return 
-     */
-    public boolean hasMaximumCaps(String message, int maxCaps, float maxPercent, String rawEmoteIndexes) {
-        // Remove Twitch emotes from the message.
-        message = getMessageWithoutEmotes(message, rawEmoteIndexes);
-        
-        boolean hasMaxCaps = false;
-        int messageLength = message.length();
-        int totalCapsCount = 1;
-
-        for (int i = 0; i < messageLength; i++) {
-            char c = message.charAt(i);
-            
-            if (Character.isUpperCase(c)) {
-                totalCapsCount++;
-                if (totalCapsCount >= maxCaps || ((totalCapsCount / messageLength) >= maxPercent)) {
-                    hasMaxCaps = true;
-                    break;
-                }
-            }
-        }
-        
-        return hasMaxCaps;
-    }
-    
-    /**
-     * Method that checks if a message is in color.
-     * 
-     * @param message
-     * @return 
-     */
-    public boolean hasColorMessage(String message) {
-        return (message.startsWith("/me ")); // The space is required for the color.
-    }
-    
-    /**
-     * Method that checks if a message is a fake purge.
-     * 
-     * @param message
-     * @return 
-     */
-    public boolean hasFakePurge(String message) {
-        boolean hasFakePurge = false;
-        
-        // Removed the colored settings, if any.
-        if (message.startsWith("/me ")) {
-            message = message.substring(4); // Remove 3 + 1 for the space.
-        }
-        
-        // Check these.
-        if (message.startsWith("<message") && message.endsWith(">")) {
-            hasFakePurge = true;
-        } else if (message.startsWith("<") && message.endsWith("deleted>")) {
-            hasFakePurge = true;
-        }
-        
-        return hasFakePurge;
     }
 }
